+++ src/extension_core/viewer_provider.ts
@@ -458,7 +458,7 @@ export class WaveformViewerProvider implements vscode.CustomReadonlyEditorProvid
     for (const signalInfo of signalList) {
       if (signalInfo.dataType && signalInfo.dataType !== 'netlist-variable') {
         const name = signalInfo.groupName;
-        this.newSignalGroup(name, groupPath, undefined);
+        this.newSignalGroup(name, groupPath, undefined, undefined);
         groupPath.push(name);
         const missing = await this.addSignalListToDocument(signalInfo.children, document, groupPath);
         missingSignals.push(...missing);
@@ -929,10 +929,11 @@ export class WaveformViewerProvider implements vscode.CustomReadonlyEditorProvid
       });
     });
 
-    let groupPath: string[] = [];
-    let index = undefined;
-    if (e.groupPath) {groupPath = e.groupPath;}
-    if (e.dropIndex) {index = e.dropIndex;}
+  let groupPath: string[] = [];
+  let index = undefined;
+  if (e.groupPath) {groupPath = e.groupPath;}
+  // Accept zero as a valid drop index
+  if (e.dropIndex || e.dropIndex === 0) {index = e.dropIndex;}
 
     if (document !== this.activeDocument) {return;}
     const metadata = netlistIdList.map(id => document.netlistIdTable[id].netlistItem);
@@ -1153,7 +1154,7 @@ export class WaveformViewerProvider implements vscode.CustomReadonlyEditorProvid
     });
   }
 
-  public newSignalGroup(name: string, groupPath: string[] | undefined, parentGroupId: number | undefined) {
+  public newSignalGroup(name: string, groupPath: string[] | undefined, parentGroupId: number | undefined, eventRowId: number | undefined) {
     if (!this.activeWebview) {return;}
     if (!this.activeDocument) {return;}
     if (!this.activeWebview.visible) {return;}
@@ -1164,6 +1165,7 @@ export class WaveformViewerProvider implements vscode.CustomReadonlyEditorProvid
       groupName: name,
       groupPath: groupPath,
       parentGroupId: parentGroupId,
+      eventRowId: eventRowId,
     });
   }